+++ Control-Box/src/main_thread.c
@@ -1,4 +1,4 @@
-#include "common.h"
+#include "common/common.h"
 #include "tasks/microSD_write_task/microSD_write_task.h"
 #include "tasks/test_gpio_task/test_gpio_task.h"
