+++ Control-Box/src/tasks/test_gpio_task/test_gpio_task.c
@@ -1,4 +1,4 @@
-#include "../../common.h"
+#include "../../common/common.h"
 #include "test_gpio_task.h"
 
 void testGpio_createTask(uint32_t pinNumber, uint32_t taskPriority, Task_Struct *taskStruct, uint8_t *taskStack){
